djangoreactfw/requirements.txt
asgiref==3.5.0
Django==4.0.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
sqlparse==0.4.2
